towhee/tests/mock_operators/pytorch_yolo_detector_operator/requirements.txt
# pip install -r requirements.txt

# Base ----------------------------------------
matplotlib>=3.2.2
numpy>=1.18.5
opencv-python>=4.1.2
Pillow>=7.1.2
PyYAML>=5.3.1
scipy>=1.4.1
torch>=1.7.0
torchvision>=0.8.1
tqdm>=4.41.0

# Logging -------------------------------------
tensorboard>=2.4.1

# Plotting ------------------------------------
pandas
seaborn>=0.11.0

# Extras --------------------------------------
thop  # FLOPs computation
